% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initDb}
\alias{initDb}
\title{initDb - Initialization of the package so the ROSMAP networks can be used with
CoExpNets}
\usage{
initDb(mandatory = F)
}
\arguments{
\item{mandatory}{}
}
\description{
initDb - Initialization of the package so the ROSMAP networks can be used with
CoExpNets
}
